+++ api_tests/src/actors/swap_factory.ts
@@ -19,6 +19,7 @@ import {
     HbitPayload,
 } from "../payload";
 import { HarnessGlobal } from "../utils";
+import { defaultExpiries, getIdentities, nowExpiries } from "./defaults";
 
 declare var global: HarnessGlobal;
 
@@ -203,46 +204,6 @@ export default class SwapFactory {
     }
 }
 
-async function getIdentities(
-    self: Actor
-): Promise<{ ethereum: string; lightning: string; bitcoin: string }> {
-    let ethereum = "0x00a329c0648769a73afac7f9381e08fb43dbea72";
-    let lightning =
-        "02ed138aaed50d2d597f6fe8d30759fd3949fe73fdf961322713f1c19e10036a06";
-    let bitcoin =
-        "02c2a8efce029526d364c2cf39d89e3cdda05e5df7b2cbfc098b4e3d02b70b5275";
-
-    try {
-        ethereum = self.wallets.ethereum.account();
-    } catch (e) {
-        self.logger.warn(
-            "Ethereum wallet not available, using static value for identity"
-        );
-    }
-
-    try {
-        lightning = await self.wallets.lightning.inner.getPubkey();
-    } catch (e) {
-        self.logger.warn(
-            "Lightning wallet not available, using static value for identity"
-        );
-    }
-
-    try {
-        bitcoin = await self.wallets.bitcoin.address();
-    } catch (e) {
-        self.logger.warn(
-            "Bitcoin wallet not available, using static value for identity"
-        );
-    }
-
-    return {
-        ethereum,
-        lightning,
-        bitcoin,
-    };
-}
-
 async function makePeer(actor: Actor): Promise<Peer> {
     return {
         peer_id: await actor.cnd.getPeerId(),
@@ -287,33 +248,5 @@ function defaultHerc20Payload(
         chain_id: 1337,
         identity,
         absolute_expiry: absoluteExpiry,
-    };
-}
-
-function defaultExpiries() {
-    const {
-        alphaAbsoluteExpiry,
-        betaAbsoluteExpiry,
-        alphaCltvExpiry,
-        betaCltvExpiry,
-    } = nowExpiries();
-
-    return {
-        alphaAbsoluteExpiry: alphaAbsoluteExpiry + 240,
-        betaAbsoluteExpiry: betaAbsoluteExpiry + 120,
-        alphaCltvExpiry,
-        betaCltvExpiry,
-    };
-}
-
-function nowExpiries() {
-    const alphaAbsoluteExpiry = Math.round(Date.now() / 1000);
-    const betaAbsoluteExpiry = Math.round(Date.now() / 1000);
-
-    return {
-        alphaAbsoluteExpiry,
-        betaAbsoluteExpiry,
-        alphaCltvExpiry: 350,
-        betaCltvExpiry: 350,
     };
 }
